requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
pillow>=10.0.0
boto3>=1.30.0